requirements.txt
torch==1.7.1
torchvision==0.8.2
mmcv_full==1.2.7
matplotlib==3.3.3
tqdm==4.46.1
omegaconf==2.0.0
numpy==1.18.5
timm==0.3.2
imageio==2.9.0
mmcv==1.3.17
Pillow==8.4.0
scikit_learn==1.0.1
